+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2021-10-28
+### Added
+- Receive thread sends a ConnectionStateAlert error on "NOTICE * :Login authentication failed"
+### Changed
+- API Changes
+    - TwitchIRC.IRC_Connect -> TwitchIRC.Connect
+    - TwitchIRC.IRC_Disconnect -> TwitchIRC.Disconnect
+    - TwitchIRC.newChatMessageEvent (UnityEvent&lt;Chatter>) -> ChatMessageEvent (event delegate(Chatter))
+    - TwitchIRC.statusEvent (UnityEvent&lt;StatusType, string, int>) -> StatusUpdateEvent (event delegate(StatusType, string, int))
+
 ## [0.2.2] - 2021-10-28
 ### Added
 - TwitchIRC.CheckSocketConnection